Moving console into utils folder
