+++ src/utils/components/Consoles/components/SerialConsole/SerialConsoleActions.tsx
