+++ src/utils/components/Consoles/components/SerialConsole/utils/serialConsole.ts
