+++ src/utils/components/Consoles/components/vnc-console/utils/VncConsoleTypes.ts
